requirements.txt
# Medical Safety Project Dependencies

# Core dependencies
openai>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
Pillow>=10.0.0

# Optional: for better terminal output
colorama>=0.4.6
